Fix resize observer bug
