+++ packages/app/src/utils/use-responsive-container.tsx
@@ -1,7 +1,6 @@
-import { ReactNode, useCallback, useEffect, useRef, useState } from 'react';
-import ResizeObserver from 'resize-observer-polyfill';
-import { useCallbackRef } from 'use-callback-ref';
+import { ReactNode, useCallback } from 'react';
 import { Box } from '~/components/base';
+import { useResizeObserver } from './use-resize-observer';
 
 /**
  * Hook returning a component which will fill to available width and height of
@@ -46,44 +45,4 @@ export function useResponsiveContainer(initialWidth: number, minHeight = 0) {
   );
 
   return { width, height, ResponsiveContainer, ref };
-}
-
-type Size = {
-  width: number;
-  height: number;
-};
-
-function useResizeObserver<T extends HTMLElement | SVGSVGElement>() {
-  const [size, setSize] = useState<Size>();
-
-  const [node, setNode] = useState<T | null>(null);
-  const ref = useCallbackRef<T>(null, (node) => setNode(node));
-
-  const observer = useRef<ResizeObserver>();
-
-  const disconnect = useCallback(() => observer.current?.disconnect(), []);
-  const connect = useCallback(
-    () =>
-      (observer.current = new ResizeObserver(
-        ([entry]: ResizeObserverEntry[]) => {
-          setSize({
-            width: Math.round(entry.contentRect.width),
-            height: Math.round(entry.contentRect.height),
-          });
-        }
-      )),
-    []
-  );
-
-  const observe = useCallback(() => {
-    connect();
-    if (node) observer.current?.observe(node);
-  }, [connect, node]);
-
-  useEffect(() => {
-    observe();
-    return () => disconnect();
-  }, [disconnect, observe]);
-
-  return { ...size, ref };
 }
